feat(ci+types): pytest+mypy in CI, hive passes mypy --strict, fleet_stats extraction
The repo's PR checks were security scanners only (GitGuardian, Socket)
-- a broken test suite would merge green. Adds a GitHub Actions
workflow running pytest and mypy on every PR and push to main.

hive/ now passes mypy --strict (enforced via [tool.mypy] in
pyproject). Most findings were missing annotations, but one was a
lying type: SQLiteTransport._watchers was declared
dict[str, list[Callable]] while actually storing
(callback, type_filter) tuples -- the annotation now tells the truth.
reputation() coerces feedback scores defensively since they arrive
from untyped JSON.

comms perf and comms fire each carried their own inline-python channel
scanner with duplicated, slightly divergent logic. Both now call
hive/fleet_stats.py (typed, stdlib-only, unit-tested), continuing the
runner_scan.py pattern of shrinking the untestable bash surface. The
fire path's tempfile + stderr-marker grep hack is gone: fire-report
and fire-json are separate clean invocations.

11 new tests (10 fleet_stats, 1 CLI perf integration). Suite: 142
passed, mypy --strict clean on 22 files.

+++ hive/coordination/reputation.py
@@ -47,4 +47,5 @@ def reputation(board: HiveBoard, agent_id: str, capability: str | None = None) -
     if total_weight == 0:
         return DEFAULT_SCORE
 
-    return sum(s * w for s, w in zip(scores, weights)) / total_weight
+    # scores come from cell data (untyped JSON); coerce defensively
+    return float(sum(float(s) * w for s, w in zip(scores, weights))) / total_weight
